Move some debugger code out of Window file
